refactor(core): move utxo-specific recovery code to utxo/
